backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
neo4j==5.18.0
openai==1.10.0
python-docx==1.1.0
PyPDF2==3.0.1
langchain==0.1.4
langchain-community==0.0.17
chromadb==0.4.22
httpx==0.26.0
tenacity==8.2.3
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
